change the structure of files
