+++ backend/package.json
@@ -23,6 +23,7 @@
     "form-data": "^4.0.2",
     "jsonwebtoken": "^9.0.2",
     "mailgun.js": "^11.1.0",
+    "morgan": "^1.10.0",
     "pg": "^8.13.1"
   },
   "devDependencies": {
